revised media cog, moved some commands to general.py
